Refact: Move AddSongAsyncTask to a separated class
