+++ app/src/main/java/com/poupa/vinylmusicplayer/discog/Discography.java
@@ -7,10 +7,7 @@ import android.view.View;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
-import com.google.android.material.snackbar.Snackbar;
-
 import com.poupa.vinylmusicplayer.App;
-import com.poupa.vinylmusicplayer.R;
 import com.poupa.vinylmusicplayer.interfaces.MusicServiceEventListener;
 import com.poupa.vinylmusicplayer.loader.ReplayGainTagExtractor;
 import com.poupa.vinylmusicplayer.loader.SongLoader;
@@ -45,7 +42,7 @@ public class Discography implements MusicServiceEventListener {
     private DB database;
     private final MemCache cache;
 
-    private View parentView;
+    View parentView;
 
     public Discography() {
         database = new DB();
@@ -146,7 +143,7 @@ public class Discography implements MusicServiceEventListener {
         new AddSongAsyncTask().execute(song);
     }
 
-    private boolean addSongImpl(@NonNull Song song, boolean cacheOnly) {
+    boolean addSongImpl(@NonNull Song song, boolean cacheOnly) {
         synchronized (cache) {
             // Race condition check: If the song has been added -> skip
             if (cache.songsById.containsKey(song.id)) {
@@ -298,75 +295,6 @@ public class Discography implements MusicServiceEventListener {
         Collection<Song> songs = database.fetchAllSongs();
         for (Song song : songs) {
             addSongImpl(song, true);
-        }
-    }
-
-    private static class AddSongAsyncTask extends AsyncTask<Song, Void, Boolean> {
-        private static int pendingCount;
-        private static int burstSongCount;
-
-        private static Snackbar progressBar;
-        private final static Discography discography = Discography.getInstance();
-
-        @Override
-        protected void onPreExecute() {
-            ++pendingCount;
-        }
-
-        @Override
-        protected Boolean doInBackground(Song... songs) {
-            boolean effectiveAdd = false;
-            for (Song song : songs) {
-                effectiveAdd = discography.addSongImpl(song, false);
-            }
-            return effectiveAdd;
-        }
-
-        @Override
-        protected void onPostExecute(Boolean result) {
-            --pendingCount;
-            if (result) {
-                ++burstSongCount;
-            }
-
-            try {
-                if (pendingCount > 0) {
-                    final String message = String.format(
-                            App.getInstance().getApplicationContext().getString(R.string.scanning_x_songs_in_progress),
-                            burstSongCount);
-                    if (progressBar == null) {
-                        progressBar = Snackbar.make(
-                                discography.parentView,
-                                message,
-                                Snackbar.LENGTH_INDEFINITE);
-                        progressBar.show();
-                    } else {
-                        progressBar.setText(message);
-                        if (!progressBar.isShownOrQueued()) {
-                            progressBar.show();
-                        }
-                    }
-                } else {
-                    if (progressBar.isShownOrQueued()) {
-                        progressBar.dismiss();
-                    }
-
-                    final String message = String.format(
-                            App.getInstance().getApplicationContext().getString(R.string.scanning_x_songs_finished),
-                            burstSongCount);
-                    Snackbar.make(
-                            Discography.getInstance().parentView,
-                            message,
-                            Snackbar.LENGTH_LONG).show();
-
-                    burstSongCount = 0;
-
-                    // Force reload the UI
-                    discography.parentView.getRootView().invalidate();
-                }
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
         }
     }
